+++ template/internal/app/routes.go
@@ -1,10 +1,10 @@
-package main
+package app
 
 import (
 	"github.com/standards-lab/go-web-sdk"
 	"github.com/standards-lab/go-web-sdk-template/template/internal/infrastructure"
 )
 
-func setRoutes(infra *infrastructure.Registry) []*web.Module {
+func routes(infra *infrastructure.Infrastructure) []*web.Module {
 	return nil
 }